Add files via upload
This is an old forum post I created (SN:Ehida) for the Star Wars Galaxies Emulation community. The game was shutdown in December of 2011, but a small community of fans/developers wanted to keep the game alive. I have been a player off and on for years with the emulation. I made this documentation for the purposes of making it more accessible to Linux users on the forum. This particular documentation dealt with POP!_OS 19 but is still a valuable resource today. It came about due to the lack of documenation for the Linux platform from the user base. There had been a handful of others that attempted to get it to work with other builds but could never fully realize the vision of getting the game that ran on a .NET framework to actually run without issues. 

The time spent troubleshooting ended up being a fun experience in patience. I wanted to share it with others in the hope that someone might find it useful.