Add files via upload
This is an example of the workflow using MojoHelpdesk. I provided documentation of how I helped a faculty member attempt to troubleshoot their issue with WebEx Meetings. Her software would not let her share her screen. I provided a simple how to guide to try to remedy the issue via search results I found on the internet. The issue ended up resolving itself, however, this is a representation of how I troubleshoot on a daily basis and how I go about providing a simple how-to guide to the end-user.